+++ src/dodal/devices/electron_analyser/vgscienta/driver_io.py
@@ -21,11 +21,9 @@ from dodal.devices.electron_analyser.abstract.types import (
     TPsuMode,
 )
 from dodal.devices.electron_analyser.util import to_kinetic_energy
-from dodal.devices.electron_analyser.vgscienta.enums import (
-    AcquisitionMode,
-    DetectorMode,
-)
+from dodal.devices.electron_analyser.vgscienta.enums import AcquisitionMode
 from dodal.devices.electron_analyser.vgscienta.region import (
+    DetectorMode,
     VGScientaRegion,
 )
