+++ tests/devices/unit_tests/electron_analyser/abstract/test_base_driver_io.py
@@ -15,7 +15,7 @@ from dodal.devices.electron_analyser.specs import (
 from dodal.devices.electron_analyser.vgscienta import (
     VGScientaAnalyserDriverIO,
 )
-from tests.devices.unit_tests.electron_analyser.helpers import (
+from tests.devices.unit_tests.electron_analyser.util import (
     TEST_SEQUENCE_REGION_NAMES,
     create_analyser_device,
 )
