--- tests/devices/unit_tests/electron_analyser/helpers/__init__.py
@@ -1,13 +1,0 @@
-from .device_factory import create_analyser_device
-from .sequence_support import (
-    TEST_SEQUENCE_REGION_NAMES,
-    assert_region_has_expected_values,
-    get_test_sequence,
-)
-
-__all__ = [
-    "create_analyser_device",
-    "TEST_SEQUENCE_REGION_NAMES",
-    "assert_region_has_expected_values",
-    "get_test_sequence",
-]
