+++ tests/devices/unit_tests/electron_analyser/vgscienta/test_region.py
@@ -13,7 +13,7 @@ from dodal.devices.electron_analyser.vgscienta.region import (
     VGScientaSequence,
 )
 from dodal.devices.i09 import LensMode, PassEnergy, PsuMode
-from tests.devices.unit_tests.electron_analyser.helpers import (
+from tests.devices.unit_tests.electron_analyser.util import (
     assert_region_has_expected_values,
     get_test_sequence,
 )
